+++ diventi/ebooks/urls.py
@@ -2,11 +2,13 @@ from django.urls import path
 from django.utils.translation import gettext_lazy as _
 
 from .views import (
-    PaperEbookView,
+    PrivatePaperEbookView,
+    PublicPaperEbookView,
 )
 
 app_name = 'ebooks'
 
 urlpatterns = [
-    path(_('paper/<slug:book_slug>/'), PaperEbookView.as_view(), name='book-detail'),
+    path(_('paper/public/<slug:book_slug>/'), PublicPaperEbookView.as_view(), name='book-detail-public'),
+    path(_('paper/<slug:book_slug>/'), PrivatePaperEbookView.as_view(), name='book-detail'),
 ]